refactor(everything): separate out structs
also do a bunch of other stuff around this
