+++ include/config/types.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+enum CmdId {
+	SERVER_ID = 0,
+	LISTEN_ID = 1,
+	SERVER_NAME_ID = 2,
+	ERROR_PAGE_ID = 3,
+	CLIENT_MAX_BODY_SIZE_ID = 4,
+	LOCATION_ID = 5,
+	LIMIT_EXCEPT_ID = 6,
+	RETURN_ID = 7,
+	ROOT_ID = 8,
+	AUTOINDEX_ID = 9,
+	INDEX_ID = 10,
+	GLOBAL_ROOT_ID = 11,
+	DATA_DIR_ID = 12,
+	UNKNOWN_ID = 404
+};
